% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{abpoly}
\alias{abpoly}
\title{Plot polynomial a + b * x + c * x^2 + d * x^3}
\usage{
abpoly(a = 0, b = 0, c = 0, d = 0, ...)
}
\arguments{
\item{a}{intercept}

\item{b}{x term}

\item{c}{x^2 term}

\item{d}{x^3 term}
}
\description{
Plot polynomial a + b * x + c * x^2 + d * x^3
}

