% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbox1.R
\name{blackbox01}
\alias{blackbox01}
\title{An unknown discrete distribution}
\usage{
blackbox01(n = 5)
}
\arguments{
\item{n}{Number of observations}
}
\value{
Numeric vector
}
\description{
An unknown discrete distribution
}

