% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{coin_flips}
\alias{coin_flips}
\title{Simulates independent coin flips}
\usage{
coin_flips(n = 1)
}
\arguments{
\item{n}{Number of coin flips}
}
\value{
a character vector of coin flips
}
\description{
Simulates independent coin flips
}

