% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{dice_rolls}
\alias{dice_rolls}
\title{Simulates independent dice rolls}
\usage{
dice_rolls(n = 1)
}
\arguments{
\item{n}{Number of die rolls}
}
\value{
a numeric vector of die rolls
}
\description{
Simulates independent dice rolls
}

