% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{ma_unif}
\alias{ma_unif}
\title{Moving average for uniformly spaced data}
\usage{
ma_unif(y, h = 1)
}
\arguments{
\item{y}{Y values}

\item{h}{Bandwidth = number of neighbors in either direction}
}
\description{
Moving average for uniformly spaced data
}

