% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{ma_unif_filt}
\alias{ma_unif_filt}
\title{Moving average for uniformly spaced data but with filter}
\usage{
ma_unif_filt(y, h = 1, filt)
}
\arguments{
\item{y}{Y values}

\item{h}{Bandwidth = number of neighbors in either direction}

\item{filt}{Logical indicating which values to filter}
}
\description{
Moving average for uniformly spaced data but with filter
}

