% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theory.R
\name{plot_discrete_distribution}
\alias{plot_discrete_distribution}
\title{Visualize a discrete distribution}
\usage{
plot_discrete_distribution(support, pmf)
}
\arguments{
\item{support}{The support of the pmf}

\item{pmf}{The probability mass function}
}
\value{
A plot
}
\description{
Visualize a discrete distribution
}

