% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbox1.R
\name{possibly_unfair_die1_hidden}
\alias{possibly_unfair_die1_hidden}
\title{An unknown discrete distribution}
\usage{
possibly_unfair_die1_hidden(n = 5)
}
\arguments{
\item{n}{Number of observations}
}
\value{
Numeric vector
}
\description{
An unknown discrete distribution
}

