% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coins.R
\name{rbernoulli}
\alias{rbernoulli}
\title{Random bernoulli trials}
\usage{
rbernoulli(n = 1, prob = 0.5)
}
\arguments{
\item{n}{number}

\item{prob}{probability of a 1}
}
\description{
Random bernoulli trials
}

