% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roll2014meta}
\alias{roll2014meta}
\title{House of representatives, metadata for first 99 rolls of 2014}
\format{An object of class \code{data.frame} with 99 rows and 6 columns.}
\source{
\url{http://http://clerk.house.gov/}
}
\usage{
roll2014meta
}
\description{
House of representatives, metadata for first 99 rolls of 2014
}
\keyword{datasets}

