% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roll2014votes}
\alias{roll2014votes}
\title{House of representatives, votes for first 99 rolls of 2014}
\format{An object of class \code{matrix} with 433 rows and 102 columns.}
\source{
\url{http://http://clerk.house.gov/}
}
\usage{
roll2014votes
}
\description{
House of representatives, votes for first 99 rolls of 2014
}
\keyword{datasets}

