% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{roll_data}
\alias{roll_data}
\title{Get data from 2014 house of representatives and print metadata}
\usage{
roll_data(rolls = 1:99, print = TRUE)
}
\arguments{
\item{rolls}{Which rolls to obtain, 1-99}

\item{print}{Whether to print metadata}
}
\value{
A dataframe
}
\description{
Get data from 2014 house of representatives and print metadata
}

