% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploratory.R
\name{sample_rows}
\alias{sample_rows}
\title{Samples rows from a table}
\usage{
sample_rows(tab, k = 5, replace = FALSE)
}
\arguments{
\item{tab}{A data fram}

\item{k}{The number of rows to sample}
}
\value{
Random subset of rows from the data frame
}
\description{
Samples rows from a table
}

