% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbox2.R
\name{taste_test_ts}
\alias{taste_test_ts}
\title{Time series data for taste test}
\usage{
taste_test_ts(xs = 1:1000)
}
\arguments{
\item{xs}{Time points, from 1 to 200}
}
\value{
Numeric vector
}
\description{
Time series data for taste test
}

