% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbox2.R
\name{taste_test_ts_hidden}
\alias{taste_test_ts_hidden}
\title{Time series data for taste test}
\usage{
taste_test_ts_hidden(xs)
}
\arguments{
\item{xs}{Time points, from 1 to 1000}
}
\value{
Numeric vector
}
\description{
Time series data for taste test
}

